Rename the 'q' alias to ':q'
It's too easy to mistype 'q<CR>' when using the 'a' alias (alias for `fasd -a`) and then your terminal closes.  I understand some people want this command, because it's similar to the vim `:q` command.  So this compromise is to make it even more similar rather than removing it outright.